backend/app/services/database/oracle_connection.py
"""
Oracle AI Vector DB Connection Module

This module handles connection to Oracle AI Vector Database.
Note: This is a placeholder for future Oracle AI Vector DB integration.
"""

import os
import logging
from typing import Optional
from dotenv import load_dotenv

logger = logging.getLogger(__name__)
load_dotenv()


class OracleConnection:
    """Manages Oracle AI Vector DB connection and operations."""
    
    def __init__(self):
        """Initialize Oracle connection with environment variables."""
        self.user = os.getenv("ORACLE_USER", "")
        self.password = os.getenv("ORACLE_PASSWORD", "")
        self.dsn = os.getenv("ORACLE_DSN", "")
        self.connection: Optional[object] = None
    
    def connect(self) -> bool:
        """Establish connection to Oracle AI Vector DB."""
        try:
            import oracledb
            
            self.connection = oracledb.connect(
                user=self.user,
                password=self.password,
                dsn=self.dsn
            )
            logger.info("Successfully connected to Oracle AI Vector DB")
            return True
        except ImportError:
            logger.error("oracledb package not installed.")
            logger.error("  Install with: conda activate orion && conda env update -f environment.yml")
            return False
        except Exception as e:
            logger.error(f"Failed to connect to Oracle AI Vector DB: {e}")
            return False
    
    def close(self):
        """Close the Oracle connection."""
        if self.connection:
            self.connection.close()
            logger.info("Oracle connection closed")
    
    def test_connection(self):
        """Test the database connection."""
        if not self.connection:
            logger.error("Connection not established")
            return False
        
        try:
            cursor = self.connection.cursor()
            cursor.execute("SELECT 1 FROM DUAL")
            result = cursor.fetchone()
            cursor.close()
            
            if result:
                logger.info("Oracle AI Vector DB connection test successful")
                return True
            return False
        except Exception as e:
            logger.error(f"Connection test failed: {e}")
            return False


def main():
    """Test the Oracle connection."""
    print("=" * 50)
    print("Oracle AI Vector DB Connection Test")
    print("=" * 50)
    
    conn = OracleConnection()
    
    if conn.connect():
        conn.test_connection()
        conn.close()
    else:
        print("\nPlease check your .env file and ensure Oracle AI Vector DB credentials are correct.")


if __name__ == "__main__":
    main()

