backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "orion-backend"
version = "0.1.0"
description = "Orion Document Management System Backend"
authors = ["Orion Team"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
neo4j = "^5.15.0"
langchain = "^0.1.0"
langchain-community = "^0.0.20"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
# Oracle DB and others can be added as needed
# oracledb = "^2.0.0" 

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
